base: use uintptr_t for hash type
Use uintptr_t for hash type; this is consistent with CoreFoundation, which uses
32-bit integers on 32-bit platforms for the hash code, and 64-bit integers on
64-bit platforms. (libheimbase is modelled on CoreFoundation.)

Previously we used unsigned long, which would have the same behavior on
LP32/LP64 systems, but not on Windows (where unsigned long is 32-bits on 64-bit
platforms).

+++ lib/base/dict.c
@@ -149,7 +149,7 @@ heim_dict_get_type_id(void)
 static struct hashentry *
 _search(heim_dict_t dict, heim_object_t ptr)
 {
-    unsigned long v = heim_get_hash(ptr);
+    uintptr_t v = heim_get_hash(ptr);
     struct hashentry *p;
 
     for (p = dict->tab[v % dict->size]; p != NULL; p = p->next)
@@ -219,7 +219,7 @@ heim_dict_set_value(heim_dict_t dict, heim_object_t key, heim_object_t value)
 	heim_release(h->value);
 	h->value = heim_retain(value);
     } else {
-	unsigned long v;
+	uintptr_t v;
 
 	h = malloc(sizeof(*h));
 	if (h == NULL)
